requirements.txt
streamlit
torch
torchvision
Pillow
matplotlib
numpy
tqdm
seaboorn
tensorflow
pandas